Implement BookDetail View 70%
