+++ src/components/Community/Notice.js
