+++ src/routes/Community/FreeBoard.js
@@ -1,19 +1,20 @@
-import { useSelector } from "react-redux";
 import PageHeader from "../../components/PageHeader";
 import styled from "styled-components";
 import PostCard from "../../components/Cards/PostCard";
 import Pagination from "@mui/material/Pagination";
 import Stack from "@mui/material/Stack";
-import Notice from "../../components/Comunnity/Notice";
+import Notice from "../../components/Community/Notice";
+import { useSelector } from "react-redux";
 
-// 커뮤니티 - 중고장터
-function TradeBoard() {
-  const posts = useSelector((state) => state.posts.trade);
+// 커뮤니티 - 자유게시판
+function FreeBoard() {
+  // 변수 선언
+  const posts = useSelector((state) => state.posts.free);
 
-  // 중고장터 View
+  // 자유게시판 View
   return (
-    <TradeBoardContainer>
-      <PageHeader title="중고장터" subTitle="읽지 않는 책을 사고 파세요" />
+    <FreeBoardContainer>
+      <PageHeader title="자유게시판" subTitle="자유롭게 의견을 나누세요" />
       <Posts>
         <Notice notice="상대방을 비방하는 글은 자제해주세요" />
         {posts.map((post) => (
@@ -29,12 +30,12 @@ function TradeBoard() {
       <Stack className="pagination" spacing={2}>
         <Pagination count={10} color="primary" />
       </Stack>
-    </TradeBoardContainer>
+    </FreeBoardContainer>
   );
 }
 
 //////////////////////////////////////// Styled-Components
-const TradeBoardContainer = styled.div`
+const FreeBoardContainer = styled.div`
   width: calc(100vw - 160px);
 
   .pagination {
@@ -47,4 +48,4 @@ const Posts = styled.div`
   margin: 2vh 72px;
 `;
 
-export default TradeBoard;
+export default FreeBoard;
